HouseMarkerResonse 자료형변경(Long), 최소/최대-월세/보증금 값 Null 방지(0으로 대체)

+++ backend_workspace/src/main/java/org/team4p/woorizip/house/dto/response/HouseMarkerResponse.java
@@ -16,12 +16,12 @@ public class HouseMarkerResponse {
 	private String houseName;
 	private double houseLat;
 	private double houseLng;
-	private int minMonthly;
-	private int maxMonthly;
-	private int minDeposit;
-	private int maxDeposit;
+	private long minMonthly;
+	private long maxMonthly;
+	private long minDeposit;
+	private long maxDeposit;
 	
-	public HouseMarkerResponse(HouseEntity houseEntity, int minDeposit, int maxDeposit, int minMonthly, int maxMonthly) {
+	public HouseMarkerResponse(HouseEntity houseEntity, long minDeposit, long maxDeposit, long minMonthly, long maxMonthly) {
 		super();
 		this.houseNo = houseEntity.getHouseNo();
 		this.houseName = houseEntity.getHouseName();
